+++ lib/kitchen/directions/bake_chapter_references/v1.rb
@@ -12,24 +12,38 @@ module Kitchen::Directions::BakeChapterReferences
 
       chapter.references.search('h3').trash
 
+      bake_page_references(page: chapter.introduction_page)
+
       chapter.non_introduction_pages.each do |page|
-        references = page.references
-        next if references.none?
-
-        references.search('h3').trash
-        title = Kitchen::Directions::EocSectionTitleLinkSnippet.v1(page: page)
-
-        references.each do |reference|
-          reference.prepend(child: title)
-        end
+        bake_page_references(page: page)
       end
 
       @content = chapter.pages.references.cut.paste
-
-      @in_composite_chapter = false
-
       chapter.append(child: render(file:
         '../../templates/eoc_section_title_template.xhtml.erb'))
+    end
+
+    def bake_page_references(page:)
+      return if page.nil?
+
+      references = page.references
+      return if references.none?
+
+      title = if page.is_introduction?
+                <<~HTML
+                  <a href="##{page.title.id}">
+                    <h3 data-type="document-title" id="#{page.title.copied_id}">
+                      <span class="os-text" data-type="" itemprop="">#{page.title_text}</span>
+                    </h3>
+                  </a>
+                HTML
+              else
+                Kitchen::Directions::EocSectionTitleLinkSnippet.v1(page: page)
+              end
+
+      references.each do |reference|
+        reference.prepend(child: title)
+      end
     end
   end
 end