+++ spec/directions/bake_chapter_references/v1_spec.rb
@@ -13,11 +13,25 @@ RSpec.describe Kitchen::Directions::BakeChapterReferences::V1 do
   let(:chapter) do
     chapter_element(
       <<~HTML
-        <div data-type="page">
-          <h1 data-type='document-title'>The Metaphysics of the Disc</h1>
+        <div data-type="page" class="introduction">
+          <h1 data-type='document-title'>Introduction to Sociology</h1>
           <section data-depth="1" id="1" class="reference">
             <h3 data-type="title">References</h3>
-            <p>Prattchett, Terry: The Color of Magic</p>
+            <p>Elias, Norbert. 1978. What Is Sociology? New York: Columbia University Press.</p>
+          </section>
+        </div>
+        <div data-type="page">
+          <h1 data-type='document-title'>What Is Sociology?</h1>
+          <section data-depth="1" id="2" class="reference">
+            <h3 data-type="title">References</h3>
+            <p>Abercrombie, Nicholas, Stephen Hill, and Bryan S. Turner. 2000. The Penguin Dictionary of Sociology. London: Penguin.</p>
+          </section>
+        </div>
+        <div data-type="page">
+        <h1 data-type='document-title'>The History of Sociology</h1>
+          <section data-depth="1" id="3" class="reference">
+            <h3 data-type="title">References</h3>
+            <p>Kierns, N. (2010). Ashley’s Alliance, unpublished presentation. Ohio State University.</p>
           </section>
         </div>
       HTML
@@ -29,8 +43,14 @@ RSpec.describe Kitchen::Directions::BakeChapterReferences::V1 do
     expect(chapter).to match_normalized_html(
       <<~HTML
         <div data-type="chapter">
+          <div data-type="page" class="introduction">
+            <h1 data-type='document-title'>Introduction to Sociology</h1>
+          </div>
           <div data-type="page">
-            <h1 data-type="document-title">The Metaphysics of the Disc</h1>
+          <h1 data-type="document-title">What Is Sociology?</h1>
+          </div>
+          <div data-type="page">
+            <h1 data-type="document-title">The History of Sociology</h1>
           </div>
           <div class="os-eoc os-references-container" data-type="composite-page" data-uuid-key=".references">
             <h2 data-type="document-title">
@@ -47,16 +67,34 @@ RSpec.describe Kitchen::Directions::BakeChapterReferences::V1 do
             <section class="reference" data-depth="1" id="1">
               <a href="#">
                 <h3 data-type="document-title" id="">
-                  <span class="os-number">1.1</span>
-                  <span class="os-divider"> </span>
-                  <span class="os-text" data-type="" itemprop="">The Metaphysics of the Disc</span>
+                  <span class="os-text" data-type="" itemprop="">Introduction to Sociology</span>
                 </h3>
               </a>
-              <p>Prattchett, Terry: The Color of Magic</p>
+              <p>Elias, Norbert. 1978. What Is Sociology? New York: Columbia University Press.</p>
+            </section>
+            <section class="reference" data-depth="1" id="2">
+              <a href="#">
+                <h3 data-type="document-title" id="">
+                  <span class="os-number">1.1</span>
+                  <span class="os-divider"> </span>
+                  <span class="os-text" data-type="" itemprop="">What Is Sociology?</span>
+                </h3>
+              </a>
+              <p>Abercrombie, Nicholas, Stephen Hill, and Bryan S. Turner. 2000. The Penguin Dictionary of Sociology. London: Penguin.</p>
+            </section>
+            <section class="reference" data-depth="1" id="3">
+              <a href="#">
+                <h3 data-type="document-title" id="">
+                  <span class="os-number">1.2</span>
+                  <span class="os-divider"> </span>
+                  <span class="os-text" data-type="" itemprop="">The History of Sociology</span>
+                </h3>
+              </a>
+              <p>Kierns, N. (2010). Ashley’s Alliance, unpublished presentation. Ohio State University.</p>
             </section>
           </div>
         </div>
-      HTML
+    HTML
     )
   end
 end